use numeric instead of bigint

+++ apps/backend/src/db/ch/schema.sql
@@ -1,10 +1,10 @@
 CREATE TABLE IF NOT EXISTS exchange.candles (
     market_id TEXT NOT NULL,
     timestamp TIMESTAMP NOT NULL,
-    open BIGINT NOT NULL,
-    high BIGINT NOT NULL,
-    low BIGINT NOT NULL,
-    close BIGINT NOT NULL,
-    volume BIGINT NOT NULL,
+    open NUMERIC NOT NULL,
+    high NUMERIC NOT NULL,
+    low NUMERIC NOT NULL,
+    close NUMERIC NOT NULL,
+    volume NUMERIC NOT NULL,
     PRIMARY KEY (market_id, timestamp)
 );
